move logger.h and constant.h to Common folder
