+++ ProcSDF/GUI/Nodes/Node.h
@@ -3,7 +3,7 @@
 #include <tuple>
 #include <vector>
 #include <map>
-#include "Constants/constant.h"
+#include "Common/constant.h"
 class TransformNode;
 class Node {
 public:
