+++ ProcSDF/Rendering/ShaderGenerator.cpp
@@ -5,7 +5,7 @@
 #include <algorithm>
 #include "Rendering/Renderer.h"
 #include "Rendering/ShaderGenerator.h"
-#include "Constants/constant.h"
+#include "Common/constant.h"
 #include "GUI/NodeGraph.h"
 void ShaderGenerator::computeAndSetObjectCount() {
